etc: emotion.d.ts 파일 분리
